Fix test case naming, Whoops
