+++ osu.Framework.Tests/Visual/OML/TestSceneOmlText.cs
@@ -3,20 +3,17 @@
 
 using NUnit.Framework;
 using osu.Framework.OML;
-using osu.Framework.Testing;
 
 namespace osu.Framework.Tests.Visual.OML
 {
-    public class OmlBox : TestScene
+    public class TestSceneOmlText : FrameworkTestScene
     {
         private const string test_data = "<oml>" +
-                                         "    <Box width=\"250\" height=\"250\" colour=\"#FF0000\"             position=\"0,0\"    />" +
-                                         "    <Box width=\"250\" height=\"250\" colour=\"green\"               position=\"100,100\"/>" +
-                                         "    <Box width=\"250\" height=\"250\" colour=\"rgba(0, 0, 255, .4)\" position=\"200,200\"/>" +
+                                         "    <text height=\"250\" width=\"250\" fontSize=\"16\">welcome to osu!</text>" +
                                          "</oml>";
 
         [Test]
-        public void TestBox()
+        public void TestText()
         {
             AddStep("Parse OML", performParse);
         }
